fixed only create content in main.rb
